refactor(agent): move worktree-* cluster into src/agent/worktree/
Move the 21-file worktree sweep/registry/occupancy cluster from the
flat src/agent/ root into src/agent/worktree/. This is the densest
prefix cluster in the directory (11 source + 10 test files) and the
single highest-value move for reducing ls noise.

What moved:
- worktree-sweep.ts (+ .test, .reconsider, valve)
- worktree-root-registry.ts (+ .test, liveness)
- worktree-root-path.ts
- worktree-occupancy.ts (+ .test)
- worktree-ignored-probe.ts (+ .test)
- worktree-ignored-patterns.ts (+ .test)
- worktree-orphan-guard.ts (+ .test)
- worktree-read-root.ts (+ .test)

What stayed:
- worktree.ts (farm module, different domain, 4 cross-layer importers)
- worktree-managed.ts family (tools/handlers/, already in a subdir)
- worktree-prune-task.ts family (daemon/, already in a subdir)

All import paths updated across src/agent/, src/cli/, including
vi.mock() and dynamic import() sites. Baselines regenerated.

Verified: pnpm lint, 19475 tests pass, filesize/funcsize/module-state/
pins gates all green.
