+++ .filesize-baseline.json
@@ -18,7 +18,7 @@
     "src/agent/tools/subagent-executor.ts": { "loc": 482, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
     "src/agent/trace/events.ts": { "loc": 426, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
     "src/agent/trace/receipt.ts": { "loc": 412, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
-    "src/agent/worktree-sweep.ts": { "loc": 500, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
+    "src/agent/worktree/worktree-sweep.ts": { "loc": 500, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
     "src/cli/_lib/testing/virtual-screen.ts": { "loc": 422, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
     "src/cli/commands/chat.ts": { "loc": 584, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
     "src/cli/commands/daemon.ts": { "loc": 399, "reason": "legacy: predates the ceiling gate; pending concern extraction" },
