+++ src/agent/daemon/worktree-prune-task.ts
@@ -4,9 +4,9 @@ import { promisify } from 'node:util';
 import { env } from '../../config/env.js';
 import { redactInlineSecrets } from '../session/prompt-dump.js';
 import type { TelemetryRecord, TelemetryTrigger } from './scheduler.js';
-import { runSweep } from '../worktree-sweep.js';
-import type { ExecFileFn, SweepResult } from '../worktree-sweep.js';
-import { sweepRootSet } from '../worktree-root-registry.js';
+import { runSweep } from '../worktree/worktree-sweep.js';
+import type { ExecFileFn, SweepResult } from '../worktree/worktree-sweep.js';
+import { sweepRootSet } from '../worktree/worktree-root-registry.js';
 import { summarizeRootFailures } from './root-failure-summary.js';
 import { countPrunable, formatWorktreePruneSummary } from './worktree-prune-summary.js';
 import { debugLog } from '../../utils/debug.js';
