+++ src/agent/subagent.ts
@@ -31,8 +31,8 @@ import { dispatchSubagentStart } from './subagent-hooks.js';
 import type { AbortOrigin, TraceSink } from './trace/index.js';
 import type { Surface } from './awareness/types.js';
 import { getCurrentSink } from './_lib/skill-sink-channel.js';
-import { touchWorktreeOccupancy, startWorktreeOccupancyHeartbeat } from './worktree-occupancy.js';
-import { resolveWorktreeMainRoot } from './worktree-read-root.js';
+import { touchWorktreeOccupancy, startWorktreeOccupancyHeartbeat } from './worktree/worktree-occupancy.js';
+import { resolveWorktreeMainRoot } from './worktree/worktree-read-root.js';
 import { type ReadScopeInputs } from './subagent-read-scope.js';
 import { resolveReadScope, composeWriteRoots } from './subagent/resolve-fork-scope.js';
 import { providerForModel, type BundledProviderName } from './providers/index.js';
