+++ src/agent/worktree/worktree-ignored-patterns.test.ts
