+++ src/agent/worktree/worktree-ignored-patterns.ts
