+++ src/agent/worktree/worktree-ignored-probe.test.ts
