+++ src/agent/worktree/worktree-ignored-probe.ts
