+++ src/agent/worktree/worktree-occupancy.test.ts
