+++ src/agent/worktree/worktree-occupancy.ts
