+++ src/agent/worktree/worktree-orphan-guard.test.ts
