+++ src/agent/worktree/worktree-orphan-guard.ts
