+++ src/agent/worktree/worktree-read-root.test.ts
