+++ src/agent/worktree/worktree-root-path.ts
