+++ src/agent/worktree/worktree-root-registry-liveness.test.ts
